requirements.txt
pysolarmanv5
prometheus_client
python-dotenv
